feat: added tbb and refactored project structure
