+++ .gitignore
@@ -1,1 +1,3 @@
-./build ./logs
+build/
+logs/
+.cache/
